Addressing Luca's PR review comments.
* Moved photolysis table free functions to impl/gas_phase_chemistry.cpp
* Replaced MAX_FIELD_NAME_LEN macro with a constexpr function
* Undefed DECLARE_PROG_TRANSFER_CONSTANTS macro after use
* Removed source file for uncoupled mam4 test
