+++ components/eamxx/src/physics/mam/mam_coupling.hpp
@@ -115,7 +115,10 @@ const char* gas_species_name(const int gas_id) {
 // here we provide storage for names of fields generated by the functions below
 namespace {
 
-#define MAX_FIELD_NAME_LEN 128
+KOKKOS_INLINE_FUNCTION
+constexpr int max_field_name_len() {
+  return 128;
+}
 
 KOKKOS_INLINE_FUNCTION
 size_t gpu_strlen(const char* s) {
@@ -151,31 +154,31 @@ void concat_3_strings(const char *s1, const char *s2, const char *s3, char *conc
 
 KOKKOS_INLINE_FUNCTION
 char* int_aero_nmr_names(int mode) {
-  static char int_aero_nmr_names_[num_aero_modes()][MAX_FIELD_NAME_LEN] = {};
+  static char int_aero_nmr_names_[num_aero_modes()][max_field_name_len()] = {};
   return int_aero_nmr_names_[mode];
 }
 
 KOKKOS_INLINE_FUNCTION
 char* cld_aero_nmr_names(int mode) {
-  static char cld_aero_nmr_names_[num_aero_modes()][MAX_FIELD_NAME_LEN] = {};
+  static char cld_aero_nmr_names_[num_aero_modes()][max_field_name_len()] = {};
   return cld_aero_nmr_names_[mode];
 }
 
 KOKKOS_INLINE_FUNCTION
 char* int_aero_mmr_names(int mode, int species) {
-  static char int_aero_mmr_names_[num_aero_modes()][num_aero_species()][MAX_FIELD_NAME_LEN] = {};
+  static char int_aero_mmr_names_[num_aero_modes()][num_aero_species()][max_field_name_len()] = {};
   return int_aero_mmr_names_[mode][species];
 }
 
 KOKKOS_INLINE_FUNCTION
 char* cld_aero_mmr_names(int mode, int species) {
-  static char cld_aero_mmr_names_[num_aero_modes()][num_aero_species()][MAX_FIELD_NAME_LEN] = {};
+  static char cld_aero_mmr_names_[num_aero_modes()][num_aero_species()][max_field_name_len()] = {};
   return cld_aero_mmr_names_[mode][species];
 }
 
 KOKKOS_INLINE_FUNCTION
 char* gas_mmr_names(int gas_id) {
-  static char gas_mmr_names_[num_aero_gases()][MAX_FIELD_NAME_LEN] = {};
+  static char gas_mmr_names_[num_aero_gases()][max_field_name_len()] = {};
   return gas_mmr_names_[gas_id];
 }
 
@@ -893,6 +896,8 @@ void transfer_work_arrays_to_prognostics(const Real q[gas_pcnst()],
     }
   }
 }
+
+#undef DECLARE_PROG_TRANSFER_CONSTANTS
 
 } // namespace scream::mam_coupling
